wpt_babel/content-security-policy/navigate-to/link-click-blocked.sub.html

<!DOCTYPE html>
<head>
 <script src="/resources/testharness.js">
 </script>
 <script src="/resources/testharnessreport.js">
 </script>
</head>
<body>
 <script>
  "use strict";

var t = async_test("Test that the child iframe navigation is not allowed");
window.onmessage = t.step_func_done(function (e) {
  assert_equals(e.data.result, 'fail');
  assert_equals(e.data.violatedDirective, 'navigate-to');
});
 </script>
 <iframe src="support/link_click_navigation.sub.html?csp=navigate-to%20%27none%27&amp;report_id={{$id:uuid()}}&amp;target=post_message_to_frame_owner.html">
 </iframe>
 <script async="" defer="" src="../support/checkReport.sub.js?reportField=violated-directive&amp;reportValue=navigate-to%20%27self%27&amp;reportID={{$id}}">
 </script>
</body>
